Email checker/requirements.txt
Flask
pandas
openpyxl
XlsxWriter
dnspython